+++ CHANGELOG.md
@@ -1,1 +1,50 @@
-## 
+## 0.1.0
+
+- Adds difficult terrain to the following spells
+  - 500 Toads
+  - Antlion Trap
+  - Ash Cloud
+  - Blast Of The Bellows
+  - Bramble Bush
+  - Burrow Ward
+  - Cave Fangs
+  - Coral Eruption
+  - Corrosive Muck
+  - Cyclone Rondo
+  - Earthworks
+  - Entangling Flora
+  - Etheric Shards
+  - Field Of Razors
+  - Flame Vortex
+  - Freezing Rain
+  - Gluttonous Growth
+  - Howling Blizzard
+  - Ice Storm
+  - Mud Pit
+  - Mushroom Patch
+  - Natures Reprisal
+  - Oneiric Mire
+  - Phantom Crowd
+  - Pillar Of Water
+  - Punishing Winds
+  - Rainbow Fumarole
+  - Rime Slick
+  - Roses Thorns
+  - Sawtooth Terrain
+  - Scatter Scree
+  - Sea Of Thought
+  - Shifting Sand
+  - Solid Fog
+  - Spike Stones
+  - Stifling Stillness
+  - Swamp Of Sloth
+  - Swampcall
+  - Telekinetic Bombardment
+  - Untwisting Iron Roots
+  - Verdant Sprout
+  - Volcanic Eruption
+  - Warped Terrain
+  - Web
+  - Whirlpool
+  - Whirlwind
+  - Wrathful Storm